+++ pytket/binders/circuit/Circuit/main.cpp
@@ -544,19 +544,19 @@ void init_circuit(py::module &m) {
           "element of `types`",
           py::arg("types"))
       .def(
-          "depth_by_2q_gates", &Circuit::depth_by_2q_gates,
+          "depth_2q", &Circuit::depth_2q,
           "Returns the number of vertices in the longest path through the "
-          "sub-DAG consisting of vertices representing 2-qubit quantum gates."
-          "Note that box types and conditionals are not included."
+          "sub-DAG consisting of vertices with 2 quantum wires,"
+          "excluding vertices representing barrier operations."
           "\n\n>>> c = Circuit(3)"
           "\n>>> c.CZ(0,1)"
           "\n>>> c.Z(0)"
           "\n>>> c.Z(1)"
           "\n>>> c.ZZMax(1,2)"
           "\n>>> c.CX(1,2)"
-          "\n>>> c.depth_by_2q_gates()"
+          "\n>>> c.depth_2q()"
           "\n3"
-          "\n:return: the circuit depth with respect to 2-qubit quantum gates.")
+          "\n:return: the circuit depth with respect to 2-qubit operations.")
       .def(
           "_to_graphviz_file", &Circuit::to_graphviz_file,
           "Saves a visualisation of a circuit's DAG to a \".dot\" file",