+++ tket/include/tket/Circuit/Circuit.hpp
@@ -1185,14 +1185,15 @@ class Circuit {
   unsigned depth_by_types(const OpTypeSet &_types) const;
 
   /**
-   * Depth of circuit restricting to 2-qubit quantum gates.
+   * Depth of circuit restricting to 2-qubit operations, excluding barriers.
    *
-   * This is the number of vertices in the longest path through the sub-DAG
-   * consisting of vertices representing 2-qubit operations.
+   * This is the number of vertices in the longest path through the
+   * sub-DAG consisting of vertices with 2 quantum wires,
+   * excluding vertices representing barrier operations.
    *
    * @return depth
    */
-  unsigned depth_by_2q_gates() const;
+  unsigned depth_2q() const;
 
   std::map<Vertex, unit_set_t> vertex_unit_map() const;
   std::map<Vertex, unsigned> vertex_depth_map() const;